refactor: CommentItem 관련 파일 이름 변경
